chore: implement BaseDrawer component; separate CampaignFilters component based on viewport
