+++ campaign-launcher/client/src/components/CampaignsFilters/index.tsx
@@ -1,68 +1,14 @@
-import { type FC, type SubmitEvent, useState } from 'react';
+import { type FC, useState } from 'react';
 
-import CheckIcon from '@mui/icons-material/Check';
-import CloseIcon from '@mui/icons-material/Close';
-import {
-  Box,
-  Button,
-  Checkbox,
-  Divider,
-  Drawer,
-  FormControlLabel,
-  IconButton,
-  Radio,
-  Stack,
-  styled,
-  Typography,
-} from '@mui/material';
+import { IconButton, styled } from '@mui/material';
 
+import ResponsiveOverlay from '@/components/ResponsiveOverlay';
 import { FilterIcon } from '@/icons';
-import { useExchangesContext } from '@/providers/ExchangesProvider';
-import { config as wagmiConfig } from '@/providers/WagmiProvider';
-import { CampaignType } from '@/types';
-import { mapTypeToLabel } from '@/utils';
+import { type CampaignType } from '@/types';
 
-const controlSlotProps = {
-  root: {
-    sx: {
-      ml: 1,
-      py: 0,
-      px: 0.5,
-    },
-  },
-};
+import CampaignsFiltersContent from './Content';
 
-const labelSlotProps = {
-  typography: {
-    color: 'white',
-  },
-};
-
-const CheckboxIcon = () => (
-  <Box
-    width={20}
-    height={20}
-    borderRadius="4px"
-    border="1.5px solid #6d6d6d"
-    bgcolor="transparent"
-  />
-);
-
-const CheckboxCheckedIcon = () => (
-  <Box
-    display="flex"
-    alignItems="center"
-    justifyContent="center"
-    width={20}
-    height={20}
-    borderRadius="4px"
-    bgcolor="error.main"
-  >
-    <CheckIcon sx={{ color: '#ffffff', fontSize: 16 }} />
-  </Box>
-);
-
-const FiltersCountStyled = styled(Box)({
+const FiltersCountStyled = styled('div')({
   display: 'flex',
   alignItems: 'center',
   justifyContent: 'center',
@@ -83,18 +29,6 @@ export type CampaignsFiltersSelection = {
   network: number;
 };
 
-const campaignTypeOptions = [...Object.values(CampaignType)].map(
-  (campaignType) => ({
-    value: campaignType,
-    label: mapTypeToLabel(campaignType),
-  })
-);
-
-const networkOptions = wagmiConfig.chains.map((chain) => ({
-  value: chain.id,
-  label: chain.name,
-}));
-
 type Props = {
   appliedFilters: CampaignsFiltersSelection;
   handleApplyFilters: (filters: CampaignsFiltersSelection) => void;
@@ -106,321 +40,63 @@ const CampaignsFilters: FC<Props> = ({
   handleApplyFilters,
   isDisabled,
 }) => {
-  const [isDrawerOpen, setIsDrawerOpen] = useState(false);
+  const [isFiltersOpen, setIsFiltersOpen] = useState(false);
   const [filtersCount, setFiltersCount] = useState(0);
-  const [draftFilters, setDraftFilters] =
-    useState<CampaignsFiltersSelection>(appliedFilters);
-
-  const { exchanges = [] } = useExchangesContext();
-
-  const exchangeOptions = exchanges.map(({ name, display_name }) => ({
-    value: name,
-    label: display_name,
-  }));
-
-  const isAllCampaignTypesSelected = draftFilters.campaignTypes.includes('');
-  const isAllExchangesSelected = draftFilters.exchanges.includes('');
-
-  const toggleAll = (section: keyof CampaignsFiltersSelection) => {
-    const currentValues = draftFilters[section] as string[];
-    const isAllSelected = currentValues.includes('');
-
-    setDraftFilters((previous) => ({
-      ...previous,
-      [section]: isAllSelected ? [] : [''],
-    }));
-  };
-
-  const toggleOption = (
-    section: keyof CampaignsFiltersSelection,
-    value: string,
-    options: string[]
-  ) => {
-    const currentValues = draftFilters[section] as string[];
-    let nextValues: string[];
-
-    if (currentValues.includes('')) {
-      nextValues = options.filter((option) => option !== value);
-    } else if (currentValues.includes(value)) {
-      nextValues = currentValues.filter(
-        (selectedValue) => selectedValue !== value
-      );
-    } else {
-      nextValues = [...currentValues, value];
-    }
-
-    setDraftFilters((previous) => ({
-      ...previous,
-      [section]: nextValues.length === options.length ? [''] : nextValues,
-    }));
-  };
-
-  const selectNetwork = (networkId: number) => {
-    setDraftFilters((previous) => ({
-      ...previous,
-      network: networkId,
-    }));
-  };
 
   const handleOpen = () => {
-    setDraftFilters(appliedFilters);
-    setIsDrawerOpen(true);
+    setIsFiltersOpen(true);
   };
 
   const handleClose = () => {
-    setDraftFilters(appliedFilters);
-    setIsDrawerOpen(false);
-  };
-
-  const handleSubmit = (event: SubmitEvent<HTMLFormElement>) => {
-    event.preventDefault();
-
-    let _filtersCount = 0;
-
-    if (draftFilters.campaignTypes.includes('')) {
-      _filtersCount += campaignTypeOptions.length;
-    } else {
-      _filtersCount += draftFilters.campaignTypes.length;
-    }
-
-    if (draftFilters.exchanges.includes('')) {
-      _filtersCount += exchangeOptions.length;
-    } else {
-      _filtersCount += draftFilters.exchanges.length;
-    }
-
-    handleApplyFilters(draftFilters);
-    setFiltersCount(_filtersCount);
-    setIsDrawerOpen(false);
+    setIsFiltersOpen(false);
   };
 
   return (
     <>
-      <Box
-        component="button"
-        display="flex"
-        alignItems="center"
-        justifyContent="center"
-        position="relative"
-        bgcolor="background.default"
-        width={42}
-        height={42}
-        border="1px solid #251d47"
-        borderRadius="100%"
-        flexShrink={0}
+      <IconButton
+        disableRipple
         disabled={isDisabled}
-        sx={{ cursor: isDisabled ? 'default' : 'pointer' }}
         onClick={handleOpen}
+        sx={{
+          p: 0,
+          position: 'relative',
+          bgcolor: 'background.default',
+          width: 42,
+          height: 42,
+          border: '1px solid #251d47',
+          borderRadius: '100%',
+        }}
       >
         <FilterIcon />
         {filtersCount > 0 && (
           <FiltersCountStyled>{filtersCount}</FiltersCountStyled>
         )}
-      </Box>
-      <Drawer
-        open={isDrawerOpen}
+      </IconButton>
+      <ResponsiveOverlay
+        open={isFiltersOpen}
         onClose={handleClose}
-        anchor="bottom"
-        slotProps={{
-          backdrop: {
-            sx: {
-              backdropFilter: 'blur(7px)',
-              background: 'rgba(0, 0, 0, 0.3)',
-            },
-          },
-          paper: {
-            elevation: 0,
-            sx: {
-              py: 2,
-              bgcolor: '#251d47',
-              borderTopLeftRadius: '16px',
-              borderTopRightRadius: '16px',
-              border: 'none',
-              maxHeight: '75dvh',
-            },
-          },
+        desktopSx={{
+          pt: 6,
+          pb: 0,
+          px: 0,
+          width: 680,
+          height: 640,
+          display: 'flex',
+          flexDirection: 'column',
+          overflow: 'hidden',
+        }}
+        mobileSx={{
+          pb: 0,
         }}
       >
-        <IconButton
-          onClick={handleClose}
-          sx={{ position: 'absolute', top: 16, right: 16, p: 0 }}
-        >
-          <CloseIcon sx={{ color: 'white' }} />
-        </IconButton>
-        <Stack
-          component="form"
-          maxHeight="100%"
-          minHeight={0}
-          onSubmit={handleSubmit}
-        >
-          <Typography variant="h6" color="white" lineHeight={1} ml={2} mb={3}>
-            Campaign Filters
-          </Typography>
-          <Stack pt={2} gap={2} minHeight={0} overflow="auto">
-            <Stack gap={1} px={2}>
-              <Typography
-                variant="caption"
-                fontSize={13}
-                fontWeight={500}
-                textTransform="uppercase"
-              >
-                Network
-              </Typography>
-              {networkOptions.map(({ label, value }) => (
-                <FormControlLabel
-                  key={value}
-                  label={label}
-                  slotProps={labelSlotProps}
-                  control={
-                    <Radio
-                      checked={draftFilters.network === value}
-                      slotProps={controlSlotProps}
-                      icon={
-                        <Box
-                          width={20}
-                          height={20}
-                          borderRadius="50%"
-                          border="1.5px solid #6d6d6d"
-                          bgcolor="transparent"
-                        />
-                      }
-                      checkedIcon={
-                        <Box
-                          display="flex"
-                          alignItems="center"
-                          justifyContent="center"
-                          width={20}
-                          height={20}
-                          borderRadius="50%"
-                          bgcolor="error.main"
-                        >
-                          <CheckIcon sx={{ color: 'white', fontSize: 16 }} />
-                        </Box>
-                      }
-                    />
-                  }
-                  onChange={() => selectNetwork(value)}
-                />
-              ))}
-            </Stack>
-            <Divider sx={{ borderColor: '#3a2e6f' }} />
-            <Stack gap={1} px={2}>
-              <Typography
-                variant="caption"
-                fontSize={13}
-                fontWeight={500}
-                textTransform="uppercase"
-              >
-                Campaign Type
-              </Typography>
-              <FormControlLabel
-                label="All"
-                control={
-                  <Checkbox
-                    checked={isAllCampaignTypesSelected}
-                    slotProps={controlSlotProps}
-                    icon={<CheckboxIcon />}
-                    checkedIcon={<CheckboxCheckedIcon />}
-                  />
-                }
-                slotProps={labelSlotProps}
-                onChange={() => toggleAll('campaignTypes')}
-              />
-              {campaignTypeOptions.map(({ label, value }) => (
-                <FormControlLabel
-                  key={value}
-                  label={label}
-                  control={
-                    <Checkbox
-                      checked={
-                        isAllCampaignTypesSelected ||
-                        draftFilters.campaignTypes.includes(value)
-                      }
-                      slotProps={controlSlotProps}
-                      icon={<CheckboxIcon />}
-                      checkedIcon={<CheckboxCheckedIcon />}
-                    />
-                  }
-                  slotProps={labelSlotProps}
-                  onChange={() =>
-                    toggleOption(
-                      'campaignTypes',
-                      value,
-                      campaignTypeOptions.map((option) => option.value)
-                    )
-                  }
-                />
-              ))}
-            </Stack>
-            <Divider sx={{ borderColor: '#3a2e6f' }} />
-            <Stack gap={1} px={2} pb={3}>
-              <Typography
-                variant="caption"
-                fontSize={13}
-                fontWeight={500}
-                textTransform="uppercase"
-              >
-                Exchanges
-              </Typography>
-              <FormControlLabel
-                label="All"
-                control={
-                  <Checkbox
-                    checked={isAllExchangesSelected}
-                    slotProps={controlSlotProps}
-                    icon={<CheckboxIcon />}
-                    checkedIcon={<CheckboxCheckedIcon />}
-                  />
-                }
-                slotProps={labelSlotProps}
-                onChange={() => toggleAll('exchanges')}
-              />
-              {exchangeOptions.map(({ label, value }) => (
-                <FormControlLabel
-                  key={value}
-                  label={label}
-                  control={
-                    <Checkbox
-                      checked={
-                        isAllExchangesSelected ||
-                        draftFilters.exchanges.includes(value)
-                      }
-                      slotProps={controlSlotProps}
-                      icon={<CheckboxIcon />}
-                      checkedIcon={<CheckboxCheckedIcon />}
-                    />
-                  }
-                  slotProps={labelSlotProps}
-                  onChange={() =>
-                    toggleOption(
-                      'exchanges',
-                      value,
-                      exchangeOptions.map((option) => option.value)
-                    )
-                  }
-                />
-              ))}
-            </Stack>
-          </Stack>
-          <Box
-            display="flex"
-            borderTop="1px solid #3a2e6f"
-            pt={3}
-            pb={1}
-            px={2}
-          >
-            <Button
-              type="submit"
-              variant="contained"
-              fullWidth
-              disableRipple
-              sx={{ bgcolor: 'error.main', color: 'white', boxShadow: 'none' }}
-            >
-              Apply filters
-            </Button>
-          </Box>
-        </Stack>
-      </Drawer>
+        <CampaignsFiltersContent
+          isOpen={isFiltersOpen}
+          appliedFilters={appliedFilters}
+          onApplyFilters={handleApplyFilters}
+          onAppliedFiltersCountChange={setFiltersCount}
+          onClose={handleClose}
+        />
+      </ResponsiveOverlay>
     </>
   );
 };
